Move JSON struct to reflect
