+++ src/main/java/at/hannibal2/skyhanni/features/inventory/EssenceShopHelper.kt
@@ -1,7 +1,7 @@
 package at.hannibal2.skyhanni.features.inventory
 
 import at.hannibal2.skyhanni.SkyHanniMod
-import at.hannibal2.skyhanni.data.jsonobjects.repo.EssenceShopUpgrade
+import at.hannibal2.skyhanni.data.jsonobjects.repo.neu.EssenceShopUpgrade
 import at.hannibal2.skyhanni.events.GuiContainerEvent
 import at.hannibal2.skyhanni.events.InventoryCloseEvent
 import at.hannibal2.skyhanni.events.InventoryFullyOpenedEvent
